Refactor: Major cleanup and project restructure after 'refactoring roadmap' meeting. Testing: Included clean folder structure as groundworks for successfull automated testing.
