--- auto-py-to-exe.txt
@@ -1,19 +1,0 @@
-altgraph==0.17.4
-auto-py-to-exe==2.41.0
-bottle==0.12.25
-bottle-websocket==0.2.9
-Eel==0.16.0
-future==0.18.3
-gevent==23.9.1
-gevent-websocket==0.10.1
-greenlet==3.0.0
-importlib-metadata==6.8.0
-macholib==1.16.3
-packaging==23.2
-pyinstaller==6.0.0
-pyinstaller-hooks-contrib==2023.9
-pyparsing==3.1.1
-whichcraft==0.6.1
-zipp==3.17.0
-zope.event==5.0
-zope.interface==6.1
