--- classes/test_classes.py
@@ -1,6 +1,0 @@
-import socket
-
-
-class local_ip:
-    def __init__(ip):
-        ip = print(socket.gethostbyname(socket.gethostname()))
