--- definitions/definition.py
@@ -1,4 +1,0 @@
-import customtkinter as ctk
-
-def change_appearance_mode_event(new_appearance_mode: str):
-        ctk.set_appearance_mode(new_appearance_mode)
