+++ requirements.txt
@@ -31,7 +31,7 @@ itsdangerous==2.1.2
 Jinja2==3.1.3
 macholib==1.16.3
 MarkupSafe==2.1.3
--e git+ssh://git@github.com/odonald/Opera-Access.git@8d55ef41347110a595ae23b558b5f95c163d427d#egg=Opera_Access
+-e git+ssh://git@github.com/odonald/Opera-Access.git@a8ed3f6ff1f944523d50aca61ed8b9c7f2edd94c#egg=Opera_Access
 outcome==1.3.0.post0
 packaging==23.2
 pillow==10.3.0
@@ -42,7 +42,7 @@ pyparsing==3.1.1
 pypng==0.20220715.0
 PySocks==1.7.1
 pytest==8.1.1
-pytest-cov==4.1.0
+pytest-cov==5.0.0
 pytest-watch==4.2.0
 qrcode==7.4.2
 redis==5.0.1
