--- tests/test_local_ip.py
@@ -1,4 +1,0 @@
-from classes.test_classes import local_ip
-
-
-local_ip()
